backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.5.1
python-multipart==0.0.6
starlette==0.27.0
jinja2==3.1.2
email-validator==2.1.0
requests==2.31.0 